The new Risk-Free MultiTask Mode (display multiTasksInOneRow)
I implemented a MultiTask Mode based on your latest master branch. This
mode is built in a way that would not lead to any conflict and can be
easily enabled by providing additional options and follow specific
constraints when providing data source. For details and example please
refer to multiTask_example.html.